+++ examples/1/views/liker.js
@@ -3,24 +3,22 @@
     var yamvc = window.yamvc,
         Liker;
 
-    Liker = yamvc.View.extend({
+    Liker = yamvc.experimental.View.extend({
         init: function () {
             Liker.Parent.init.apply(this, arguments);
-            this.bindEvents();
-        },
-        bindEvents: function () {
-            this.getModel('likes')
-                .addListener('dataCountChange', this.updateLiker.bind(this));
         },
         incrementLikes: function () {
-            this.getModel('likes')
-                .data(
-                    'count',
-                    this.getModel('likes').data('count') + 1
-                );
-        },
-        updateLiker: function () {
-            this.partialRender('.liker');
+            var model = this.getModel('likes');
+
+            model.data(
+                'count',
+                model.data('count') + 1
+            );
+
+            model.data(
+                'disabled',
+                !model.data('disabled')
+            );
         }
     });